Improve startup time for converter.
Don't cluster "uncategorized" tables - the performance improvement isn't big enough to justify the time taken.

Fix the position of categories:cid in TODO.md.

Require that at least one argument is passed to ./run.bash.

+++ run.bash
@@ -2,6 +2,12 @@
 
 set -e
 
+if [[ "$#" -eq 0 ]]; then
+	echo './run.bash requires MAKE(1) parameters.'
+	echo 'Run "./run.bash -j1" for generic "make".'
+	exit 2
+fi
+
 rm -f Makefile secret.go brute_ips
 
 echo 'package main' > secret.go
